+++ shared/src/main/scala/com/miriamlaurel/fxcore/market/Order.scala
@@ -8,8 +8,8 @@ case class Order(key: OrderKey,
                  amount: SafeDouble,
                  price: SafeDouble) extends Ordered[Order] {
 
-  require(amount.toDouble > 0)
-  require(price.toDouble > 0)
+  require(amount > 0)
+  require(price > 0)
 
   override def compare(that: Order) = if (key.side == QuoteSide.Ask) price compare that.price else that.price compare price
