+++ shared/src/main/scala/com/miriamlaurel/fxcore/portfolio/Position.scala
@@ -32,7 +32,7 @@ case class Position(primary: Monetary,
     * If position's primary amount is positive (a long position), it's secondary amount should be negative,
     * and vice versa.
     */
-  require(primary.amount.toDouble.signum != secondary.amount.toDouble.signum)
+  require(primary.amount.signum != secondary.amount.signum)
 
   /*!
   An instrument can be constructed from position's primary and secondary asset, like "EUR/USD" (a currency pair).
@@ -42,7 +42,7 @@ case class Position(primary: Monetary,
   /*!
   We can easily calculate a price for 1 unit of primary asset -- it is usually referred as position's price.
    */
-  lazy val price: SafeDouble = math.abs(secondary.amount.toDouble / primary.amount.toDouble)
+  lazy val price: SafeDouble = math.abs(secondary.amount / primary.amount)
 
   /*!
   Position can be either long (primary amount is positive) or short (negative). Short positions may be handled
@@ -54,7 +54,7 @@ case class Position(primary: Monetary,
   Position amount is the absolute value of primary asset amount, i.e. it is always positive. If amount sign
   matters (it usually does), one should use .primary.amount
    */
-  lazy val absoluteAmount: SafeDouble = math.abs(primary.amount.toDouble)
+  lazy val absoluteAmount: SafeDouble = math.abs(primary.amount)
 
   /*!
   Create a reversed position with new price and matching UUID. This can be used efficiently to close a position.
@@ -111,17 +111,17 @@ case class Position(primary: Monetary,
     val a2 = this.secondary.amount
     val b1 = that.primary.amount
     val b2 = that.secondary.amount
-    val c1: SafeDouble = (if (a1.toDouble.signum * b1.toDouble.signum == -1) math.abs(a1.toDouble) min math.abs(b1.toDouble) else 0) * a1.toDouble.signum
-    val c2: SafeDouble = if (a1.toDouble == 0) a2 else c1 * (a2 / a1)
+    val c1: SafeDouble = (if (a1.signum * b1.signum == -1) math.abs(a1) min math.abs(b1) else 0) * a1.signum
+    val c2: SafeDouble = if (a1 == 0) a2 else c1 * (a2 / a1)
     val d1: SafeDouble = -c1
-    val d2: SafeDouble = if (b1.toDouble == 0) b2 else d1 * (b2 / b1)
+    val d2: SafeDouble = if (b1 == 0) b2 else d1 * (b2 / b1)
     // e1 is always zero
     val e2: SafeDouble = c2 + d2
-    val sigma = if (math.abs(a1.toDouble) > math.abs(b1.toDouble)) -1 else 1
+    val sigma = if (math.abs(a1) > math.abs(b1)) -1 else 1
     val f1: SafeDouble = a1 + b1
-    val f2: SafeDouble = if (a1.toDouble.signum * b1.toDouble.signum == 1) a2 + b2 else if (sigma < 0) a2 - c2 else b2 - d2
+    val f2: SafeDouble = if (a1.signum * b1.signum == 1) a2 + b2 else if (sigma < 0) a2 - c2 else b2 - d2
 
-    val pos: Option[Position] = if (f1.toDouble == 0) None
+    val pos: Option[Position] = if (f1 == 0) None
     else
       Some(Position(Monetary(f1, primary.asset), Monetary(f2, secondary.asset), None, that.timestamp, UUID.randomUUID()))
     (pos, Money(e2, secondary.asset))
@@ -160,7 +160,7 @@ case class Position(primary: Monetary,
     require(oldPosition.instrument == this.instrument)
     require(oldPosition.side != this.side)
     require(oldPosition.absoluteAmount != this.absoluteAmount)
-    val closingAmount = math.min(oldPosition.absoluteAmount.toDouble,this.absoluteAmount.toDouble) * oldPosition.primary.amount.toDouble.signum
+    val closingAmount = math.min(oldPosition.absoluteAmount,this.absoluteAmount) * oldPosition.primary.amount.signum
     val closingPart = Position(oldPosition.instrument, oldPosition.price,
       closingAmount, Some(oldPosition.id), oldPosition.timestamp)
     Deal(closingPart, this.price, this.timestamp, (oldPosition merge this)._2)
